backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
